Migrate to init.lua
To learn more lua, and since lua seems to be the future for neovim
configuration.

+++ install.sh
@@ -2,5 +2,10 @@
 
 mkdir -p "${HOME}/.config/nvim"
 
-ln -sf "${PWD}/init.vim" "${HOME}/.config/nvim/init.vim"
+if [ -f "${PWD}/init.lua" ]; then
+  ln -sf "${PWD}/init.lua" "${HOME}/.config/nvim/init.lua"
+elif [ -f "${PWD}/init.vim" ]; then
+  ln -sf "${PWD}/init.vim" "${HOME}/.config/nvim/init.vim"
+fi
+
 ln -snf "${PWD}/lua" "${HOME}/.config/nvim/lua"
